🎉 v1.2.3 Added Playlist Search!
